implement allowed ranges
